+++ pkg/querier/distributor_queryable_streaming_test.go
@@ -21,6 +21,7 @@ import (
 	"github.com/grafana/mimir/pkg/querier/batch"
 	"github.com/grafana/mimir/pkg/querier/stats"
 	"github.com/grafana/mimir/pkg/storage/chunk"
+	"github.com/grafana/mimir/pkg/streamingpromql/limiting"
 	"github.com/grafana/mimir/pkg/util/limiter"
 )
 
@@ -158,8 +159,10 @@ func createTestStreamReader(batches ...[]client.QueryStreamSeriesChunks) *client
 	}
 
 	cleanup := func() {}
+	queryLimiter := limiter.NewQueryLimiter(0, 0, 0, 0, nil)
+	memoryTracker := limiting.NewMemoryConsumptionTracker(0, nil)
 
-	reader := client.NewSeriesChunksStreamReader(ctx, mockClient, "ingester", seriesCount, limiter.NewQueryLimiter(0, 0, 0, 0, nil), cleanup, log.NewNopLogger())
+	reader := client.NewSeriesChunksStreamReader(ctx, mockClient, "ingester", seriesCount, queryLimiter, memoryTracker, cleanup, log.NewNopLogger())
 	reader.StartBuffering()
 
 	return reader